+++ app/javascript/vue/components/otu/otu.vue
@@ -2,7 +2,7 @@
   <div class="otu-radial">
     <VBtn
       :title="redirect ? 'Browse OTUs' : 'OTU quick forms'"
-      :color="[emptyList ? 'create' : redirect ? 'primary' : 'radial']"
+      :color="emptyList ? 'create' : redirect ? 'primary' : 'radial'"
       circle
       @click="openApp()"
       @contextmenu.prevent="openApp(true)"