+++ app/javascript/vue/tasks/biological_associations/biological_associations_graph/components/BiologicalAssociationGraph.vue
@@ -144,6 +144,7 @@
       @update:name="
         ($event) => {
           setGraphName($event)
+          saveGraph()
           showModalGraph = false
         }
       "
@@ -212,6 +213,8 @@ import ContextMenuView from './ContextMenu/ContextMenuView.vue'
 import ContextMenuNode from './ContextMenu/ContextMenuNode.vue'
 import { makeNodeObject } from '../adapters'
 
+const emit = defineEmits('load:graph')
+
 const {
   addBiologicalRelationship,
   addCitationFor,
@@ -237,6 +240,7 @@ const {
   reverseRelation,
   save,
   saveBiologicalAssociations,
+  saveGraph,
   selectedEdges,
   selectedNodes,
   setGraphName,
@@ -385,9 +389,25 @@ function addNodeObject(obj) {
   addObject(makeNodeObject(obj))
 }
 
+async function downloadAsSvg() {
+  if (!graph.value) return
+  const text = await graph.value.exportAsSvgText()
+  const url = URL.createObjectURL(new Blob([text], { type: 'octet/stream' }))
+  const a = document.createElement('a')
+  a.href = url
+  a.download = 'network-graph.svg'
+  a.click()
+  window.URL.revokeObjectURL(url)
+}
+
+function getBiologicalRelationships() {
+  return biologicalAssociations
+}
+
 defineExpose({
   addNodeObject,
   currentNodes,
+  getBiologicalRelationships,
   isGraphUnsaved,
   loadBiologicalAssociations,
   openEdgeModal,
@@ -397,7 +417,8 @@ defineExpose({
   resetStore,
   save,
   saveBiologicalAssociations,
-  setGraph
+  setGraph,
+  downloadAsSvg
 })
 </script>